start of modularizing modals
