+++ app/modals/ServicesModal/ServicesModal.scss
@@ -1,4 +1,4 @@
-@import './constants.scss';
+@import '../../stylesheets/constants.scss';
 
 .servicesContainer {
   width: 100px;
