+++ app/modals/UserModal/UserModal.scss
@@ -1,4 +1,4 @@
-@import './constants.scss';
+@import '../../stylesheets/constants.scss';
 
 .add-container {
   @include centerModal(30%, 500px);
